Commit for nscldaq-11.0-016 after merge with branch features-nscldaq-11.0-014/

+++ main/daq/readoutgui/eventLogBundle.tcl
@@ -21,6 +21,8 @@ package provide eventLogBundle 1.0
 package provide Experiment     2.0;   # For compatibility with event builder.
 
 
+package require stageareaValidation
+
 package require Tk
 package require DAQParameters
 package require RunstateMachine
@@ -110,7 +112,12 @@ namespace eval ::EventLog {
     
     ## 
     # @var failed         - Indicates whether or not the system has failed or succeeded
-    variable failed 0 
+    variable failed 0
+    
+    ##
+    # incremented when the eventLogger exits...this can be a vwait target:
+    
+    variable eventLogEnded 0
     
     # Export the bundle interface methods
     
@@ -356,12 +363,13 @@ proc ::EventLog::_handleInput {} {
 
         if {!$::EventLog::expectingExit} {
             ::ReadoutGUIPanel::Log EventLogManager error "Unexpected event log error! $msg"
-            ::Diagnostics::Error {The event logger exited unexpectedly!!}
+            ::Diagnostics::Error {The event logger exited unexpectedly check EventLogManager tab for errors.!!}
         } else {
             ::EventLog::_finalizeRun;            # May need that if exit before wait.
         }
         set ::EventLog::loggerFd [list]
         set ::EventLog::loggerPid -1
+        incr ::EventLog::eventLogEnded
     } else {
         set line [gets $fd]
         ::ReadoutGUIPanel::Log EventLogManager output $line
@@ -512,147 +520,6 @@ proc ::EventLog::_setStatusLine repeatInterval {
     set ::EventLog::statusUpdateId \
         [after $repeatInterval [list ::EventLog::_setStatusLine $repeatInterval]]
 }
-##
-# ::EventLog::_duplicateRun
-#
-#  @return string -   
-#  @retval non empty  if the run we are about to write already exists.
-#          we're going to define 'exists' as having a run directory in the
-#          experiment view. 
-# @retval empty if there is no sign that this run had already been recorded.
-#
-proc ::EventLog::_duplicateRun {} {
-    
-    set run [::ReadoutGUIPanel::getRun]
-    
-    # Two possibilities;  If the run was properly ended, there will be a
-    # run directory in the experimnent view
-    
-    set runDirPath [::ExpFileSystem::getRunDir $run]
-    
-    # If the run was improperly ended, there could be event segments in the
-    # current directory.  We'll look for them with glob.
-    #
-    
-    set checkGlob [::ExpFileSystem::getCurrentRunDir]
-    set checkGlob [file join $checkGlob [::ExpFileSystem::genEventfileBasename $run]*.evt ]
-    
-    set eventSegments [llength [glob -nocomplain $checkGlob]]
-
-    #  Figure out the return value:
-
-    if {[file exists $runDirPath]} {
-      set message "The final run directory '$runDirPath' already exists indicating this run may already have been recorded"
-    } elseif {($eventSegments > 0)} {
-      set message "[::ExpFileSystem::getCurrentRunDir] has event segments in it for this run ($run) indicating this run may already have been recorded but not finalized"
-    } else {
-      set message ""
-    }
-    return $message
-
-}
-
-##
-#   Check whether or not the .started file lives in the 
-#   experiment/current directory
-#
-proc ::EventLog::_dotStartedExists {} {
-  set currentPath [::ExpFileSystem::getCurrentRunDir]
-  return [file exists [file join $currentPath .started]]
-}
-
-##
-#   Check whether or not the .exited file lives in the 
-#   experiment/current directory
-#
-proc ::EventLog::_dotExitedExists {} {
-  set currentPath [::ExpFileSystem::getCurrentRunDir]
-  return [file exists [file join $currentPath .exited]]
-}
-
-##
-#   Check whether or not .evt files exist in the 
-#   experiment/current directory
-#
-#   @returns boolean indicating whether there are any files ending in .evt
-#
-proc ::EventLog::_runFilesExistInCurrent {} {
-  set currentPath [::ExpFileSystem::getCurrentRunDir]
-  set evtFiles [glob -directory $currentPath -nocomplain *.evt]
-  return [expr {[llength $evtFiles] > 0} ]
-}
-
-##
-# ::EventLog::listIdentifiableProblems
-#
-# Checks for a few things:
-# 1. experiment/run# directory already exists
-# 2. experiment/current/*.evt files exist
-#
-# @returns a list of error messages
-proc ::EventLog::listIdentifiableProblems {} {
-
-  set errors [list]
-
-  # check if run directory exist!
-  set msg [EventLog::_duplicateRun]
-  if {$msg ne ""} {
-    lappend errors $msg
-  } 
-  
-  # check if experiment/current/*.evt files exist
-  if {[::EventLog::_runFilesExistInCurrent]} {
-    set msg    "EventLog error: the experiment/current directory contains run "
-    append msg "segments and needs to be cleaned."
-    lappend errors $msg
-  }
-
-  return $errors
-}
-##
-# correctFixableProblems
-#
-#  Some startup issues can be corrected:
-# 1. experiment/current/.started exists
-# 2. experiment/current/.exited exists
-#
-#  In this case these files are just deleted.
-#
-proc ::EventLog::correctFixableProblems {} {
-  # check if experiment/current/.started exists
-  if {[::EventLog::_dotStartedExists]} {
-    ::EventLog::deleteStartFile
-
-  } 
-  
-  # check if experiment/current/.exited exists
-  if {[::EventLog::_dotExitedExists]} {
-    ::EventLog::deleteExitFile
-  } 
-    
-}
-
-##
-# deleteStartFile
-#   Kill off the .started file.
-#
-proc ::EventLog::deleteStartFile {} {
-    set startFile [file join [::ExpFileSystem::getCurrentRunDir] .started]
-    file delete -force $startFile 
-    
-}
-##
-# deleteExitFile
-#   Delete the .exited file
-#
-proc ::EventLog::deleteExitFile {} {
-    
-    set exitFile [file join [::ExpFileSystem::getCurrentRunDir] .exited]
-    file delete -force $exitFile 
-    
-}
-#------------------------------------------------------------------------------
-# Actions:
 
 ##
 # ::EventLog::runStarting
@@ -674,20 +541,16 @@ proc ::EventLog::runStarting {} {
   }
 
   # Now if desired start the new run.
+  ::StageareaValidation::correctAndValidate
 
   if {[::ReadoutGUIPanel::recordData]} {
-    ::EventLog::correctFixableProblems;         # Some things can be fixed :-)
-    set errorMessages [::EventLog::listIdentifiableProblems]
-    if {[llength $errorMessages]>0} {
-      return -code error $errorMessages
-    }
     
     set startFile [file join [::ExpFileSystem::getCurrentRunDir] .started]
 
     ::EventLog::_startLogger
     ::EventLog::_waitForFile $startFile $::EventLog::startupTimeout \
                                         $::EventLog::filePollInterval
-    ::EventLog::deleteStartFile
+    ::StageareaValidation::deleteStartFile
     set ::EventLog::expectingExit 0
     ::EventLog::_setStatusLine 2000
     set ::EventLog::needFinalization 1
@@ -707,18 +570,49 @@ proc ::EventLog::runEnding {} {
     
     set startFile [file join [::ExpFileSystem::getCurrentRunDir] .started]
     set exitFile [file join [::ExpFileSystem::getCurrentRunDir] .exited]
+
     # ne is used below because the logger could be a pipeline in which case
     # ::EventLog::loggerPid will be a list of pids which freaks out ==.
     
     if {$::EventLog::loggerPid ne -1} {
         
         set ::EventLog::expectingExit 1
-        ::EventLog::_waitForFile $exitFile $::EventLog::shutdownTimeout \
-            $::EventLog::filePollInterval
-        foreach pid $::EventLog::loggerPid {
-          catch {exec kill -9 $pid}; # in case waitforfile timed out.
+        
+        #  First do a vwait for eventLogEnded after disabling the
+        #  begin/end etc. buttons.
+        #  A timeout is used in case there's a problem and the event log
+        #  never exists.
+        
+        set ui [::RunControlSingleton::getInstance]
+        $ui configure -state disabled
+        
+        set timeoutId [after \
+            [expr {$::EventLog::shutdownTimeout*1000}]    \
+            [list incr ::EventLog::eventLogEnded]         \
+        ]
+        set oldValue $::EventLog::eventLogEnded;      # so we know if there was a timeout.
+        vwait ::EventLog::eventLogEnded
+        after cancel $timeoutId
+
+        if {$oldValue == $::EventLog::eventLogEnded} {
+            
+            # Wait timed out.
+            tk_messageBox -title "EventLogger exit timeout" -icon warning -type ok \
+                -message {Timed out waiting for eventlog to exit, killing it}
+            foreach pid $::EventLog::loggerPid {
+              catch {exec kill -9 $pid}; # in case waitforfile timed out.
+            }
+        } else {
+            # Normal exit, this should fall through very quickly, now that
+            #  we know the logger exited.
+            
+            ::EventLog::_waitForFile $exitFile $::EventLog::shutdownTimeout \
+                $::EventLog::filePollInterval
+            ::EventLog::deleteExitFile
+            
         }
-        ::EventLog::deleteExitFile
+        puts "Enabling"
+        ::StageareaValidation::deleteExitFile
         set ::EventLog::loggerPid -1
         ::EventLog::_finalizeRun
         file delete -force $startFile;   # So it's not there next time!!
@@ -794,12 +688,19 @@ proc ::EventLog::enter {from to} {
 #
 proc ::EventLog::leave {from to} {
   if {($from eq "Halted") && ($to eq "Active")} {
-      ::EventLog::runStarting
+      if {[catch {::EventLog::runStarting} msg]} {
+        set ::EventLog::failed 1
+        ::ReadoutGUIPanel::Log EventLogManager error $msg
+        error $msg
+      }
       # reset the failure state
       set ::EventLog::failed 0
   }
   if {($from in [list "Active" "Paused"]) && ($to eq "Halted") } {
-    set  ::EventLog::expectingExit 1
+    if {! $::EventLog::failed} { 
+      set  ::EventLog::expectingExit 1
+    }
+    set ::EventLog::failed 0
   }
 }
